\name{print.contrastResult}
\alias{print.contrastResult}
\title{
Printing a \code{contrastResult} object
}
\description{Generic S3 function to print the results of a \code{contrastResult} object created with \code{\link{contrast}}.}
\usage{
\method{print}{contrastResult}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{contrastResult}.

}
  \item{\dots}{}
}

\value{No return value, called for printing the results of an object of class \code{contrastResult}.}

\keyword{internal}