\name{print.gformula}
\alias{print.gformula}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Printing the results of a \code{gformula} object
}
\description{Generic S3 function to print the results of an object created with \code{\link{gformula}}.}
\usage{
\method{print}{gformula}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of \code{class} `gformula'.

}
  \item{\dots}{}
}

\value{No return value, called for printing the results of an object of class \code{class} `gformula'.}


\keyword{internal}
