\name{prop}
\alias{prop}
\title{Helper function to calculate proportions of categorical variables}
\description{To be used as \code{fun} argument in \code{\link{custom.measure}}.}
\arguments{
  \item{vec}{A vector with either numeric, logical or string elements.}
  \item{categ}{A  numeric, logical or string vector of length 1, that refers to a particular category of \code{vec}.}
}
\value{A scalar numeric value of the proportion of \code{categ} values in \code{vec}}
\keyword{internal}




